✨ Feat(강의 데이터 삭제): 강의 없는 경우 처리

+++ insty-api/src/main/java/insty/domain/course/implement/CourseCleaner.java
@@ -36,13 +36,16 @@ public class CourseCleaner {
 
     public void cleanAllData(Long userId) {
         List<Long> courseIds = courseRepository.findAllIdByUserId(userId);
-        List<String> keys = new ArrayList<>();
+        if (courseIds.isEmpty()) {
+            return;
+        }
+        List<String> s3Keys = new ArrayList<>();
 
         deleteAllTag(courseIds);
-        deleteAllFile(keys, courseIds);
-        deleteAllVideo(keys, courseIds);
+        deleteAllFile(s3Keys, courseIds);
+        deleteAllVideo(s3Keys, courseIds);
         deleteAllCourse(courseIds);
-        s3FileManager.deleteAllByKeyList(keys);
+        s3FileManager.deleteAllByKeyList(s3Keys);
     }
 
     private void deleteAllTag(List<Long> courseIds) {
